firsts steps of the Milestone 3
